Crate tlb_shootdown

Support for broadcasting and handling TLB shootdown IPIs.

Statics

TLB_SHOOTDOWN_IPI_COUNT

The number of remaining cores that still need to handle the curerent TLB shootdown IPI

TLB_SHOOTDOWN_IPI_LOCK

The lock that makes sure only one set of TLB shootdown IPIs is concurrently happening

TLB_SHOOTDOWN_IPI_PAGES

The range of pages for a TLB shootdown IPI.

Functions

handle_tlb_shootdown_ipi

Handles a TLB shootdown ipi by flushing the VirtualAddresses covered by the given range of pages_to_invalidate.

init

Initializes data, functions, and structures for the TLB shootdown. TODO: redesign this, it’s weird and silly just to set one callback.

send_tlb_shootdown_ipi

Sends an IPI to all other cores (except me) to trigger a TLB flush of the given pages’ virtual addresses.